Use generic squash query for possible for squashable counts
