+++ temba/orgs/models.py
@@ -1881,24 +1881,6 @@ class ItemCount(BaseScopedCount):
 
     org = models.ForeignKey(Org, on_delete=models.PROTECT, related_name="counts", db_index=False)  # indexed below
 
-    @classmethod
-    def get_squash_query(cls, distinct_set: dict) -> tuple:
-        sql = """
-        WITH removed as (
-            DELETE FROM %(table)s WHERE "org_id" = %%s AND "scope" = %%s RETURNING "count"
-        )
-        INSERT INTO %(table)s("org_id", "scope", "count", "is_squashed")
-        SELECT %%s, %%s, s.total, TRUE FROM (
-            SELECT COALESCE(SUM("count"), 0) AS "total" FROM removed
-        ) s WHERE s.total != 0;
-        """ % {
-            "table": cls._meta.db_table
-        }
-
-        params = (distinct_set["org_id"], distinct_set["scope"]) * 2
-
-        return sql, params
-
     class Meta:
         indexes = [
             models.Index("org", OpClass("scope", name="varchar_pattern_ops"), name="orgcount_org_scope"),
