+++ temba/utils/models/__init__.py
@@ -1,3 +1,2 @@
 from .base import *  # noqa
 from .fields import JSONAsTextField, JSONField, TranslatableField  # noqa
-from .squashable import *  # noqa
